+++ puzzle_ops/models.py
@@ -80,6 +80,7 @@ class DemandRow:
     value_match: str = ""
     reference_image_url: str = ""
     reference_image_path: str = ""
+    reference_image_asset_id: str = ""
     reference_image_content_type: str = ""
     reference_image_syncable: bool = True
     generation_review_status: str = ""